lift: call cb with no results if orig fn returns undefined
This ensures that when calling lift on a function that returns undefined (and say, just throws an
error in some cases) doesn't cause async to think it returned one value
(the value undefined), since this will mess up the arguments in the waterfall.
async will put undefined as the first arg to the next function in the
waterfall.